+++ src/data/series/FloatSeries.java
@@ -4,6 +4,6 @@ package data.series;
  * Created by galafit on 1/1/18.
  */
 public interface FloatSeries {
-    public int size();
-    public float get(int index);
+    public long size();
+    public float get(long index);
 }
