+++ src/data/series/grouping/aggregation/FloatAggregateFunction.java
@@ -6,6 +6,7 @@ import data.series.FloatSeries;
  * Created by galafit on 12/10/17.
  */
 public interface FloatAggregateFunction {
-    public float group(FloatSeries series, int from, int length);
+    public float group(FloatSeries series, long from, int length);
 
 }
+
