+++ src/data/series/grouping/aggregation/FloatMax.java
@@ -7,7 +7,7 @@ import data.series.FloatSeries;
  */
 public class FloatMax implements FloatAggregateFunction {
     @Override
-    public float group(FloatSeries series, int from, int length) {
+    public float group(FloatSeries series, long from, int length) {
         float max = series.get(from);
         for (int i = from + 1; i < from + length; i++) {
             max = Math.max(max, series.get(i));
